+++ testapp/models.py
@@ -1,16 +1,11 @@
 """Models for the testapp."""
-from enum import IntEnum
 
 from django.db import models
 from ndh.models import Links, NamedModel, TimeStampedModel
-from ndh.utils import enum_to_choices
 
 
 class TestModel(Links, TimeStampedModel, NamedModel):
     """Main test model."""
-    YEARS_IN_SCHOOL = IntEnum('years_in_school', 'freshman sophomore junior senior')
-
-    year_in_school = models.IntegerField(choices=enum_to_choices(YEARS_IN_SCHOOL), default=1)
     tests = models.IntegerField(default=42)
     tests_decimal = models.DecimalField(default=3.14, max_digits=5, decimal_places=2)
     moment = models.DateTimeField()